# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Conversational Order Operating System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /public/s/{storeSlug}/p/{productSlug}/order:
    post:
      operationId: createPublicOrderByStore
      tags: [public]
      summary: Create order from a store-scoped product page
      parameters:
        - name: storeSlug
          in: path
          required: true
          schema:
            type: string
        - name: productSlug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublicOrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicOrderResult"
        "404":
          description: Store or product not found
        "422":
          description: Invalid product variant or delivery selection
...
  /stores/{storeId}/orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders for a store
      parameters:
        - name: storeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrdersPage"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create an order (from landing page)
      parameters:
        - name: storeId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /stores/{storeId}/orders/{orderId}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get an order
      parameters:
        - name: storeId
          in: path
          required: true
          schema:
            type: integer
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
    patch:
      operationId: updateOrder
      tags: [orders]
      summary: Update order status or notes
      parameters:
        - name: storeId
          in: path
          required: true
          schema:
            type: integer
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /stores/{storeId}/orders/summary:
    get:
      operationId: getOrdersSummary
      tags: [orders]
      summary: Get order counts by status (dashboard summary)
      parameters:
        - name: storeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrdersSummary"

  # ─── Customers ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    PublicOrderInput:
      type: object
      required: [customerName, customerPhone, deliveryZoneId, deliveryCommuneName, deliveryMethod]
      properties:
        customerName:
          type: string
        customerPhone:
          type: string
        customerCity:
          type: string
        customerAddress:
          type: string
        deliveryZoneId:
          type: integer
        deliveryCommuneName:
          type: string
        deliveryMethod:
          type: string
          enum: [HOME, OFFICE]
        selectedSize:
          type: string
        selectedColor:
          type: string
        quantity:
          type: integer
          minimum: 1
          maximum: 10
        notes:
          type: string
    PublicOrderResult:
      type: object
      required: [id, productName, totalPrice, deliveryFee, payableTotal, status]
      properties:
        id:
          type: integer
        productName:
          type: string
        totalPrice:
          type: number
        deliveryFee:
          type: number
        payableTotal:
          type: number
        status:
          type: string
...
    DeliveryMethod:
      type: string
      enum: [HOME, OFFICE]
...
    OrderStatus:
      type: string
      enum:
        - NEW
        - PENDING_CONFIRMATION
        - CONFIRMED
        - SHIPPED
        - DELIVERED
        - RETURNED
        - CANCELLED
        - REJECTED
    Order:
      type: object
      required: [id, storeId, landingPageId, customerName, customerPhone, customerCity, quantity, totalPrice, status, createdAt]
      properties:
        id:
          type: integer
        storeId:
          type: integer
        landingPageId:
          type: integer
        landingPageName:
          type: ["string", "null"]
        productImageUrl:
          type: ["string", "null"]
        customerId:
          type: ["integer", "null"]
        customerName:
          type: string
        customerPhone:
          type: string
        customerCity:
          type: string
        customerAddress:
          type: ["string", "null"]
        deliveryZoneId:
          type: ["integer", "null"]
        deliveryWilayaCode:
          type: ["string", "null"]
        deliveryWilayaName:
          type: ["string", "null"]
        deliveryCommuneName:
          type: ["string", "null"]
        deliveryDairaName:
          type: ["string", "null"]
        deliveryMethod:
          anyOf:
            - $ref: "#/components/schemas/DeliveryMethod"
            - type: "null"
        deliveryFee:
          type: number
        returnFee:
          type: number
        payableTotal:
          type: number
        selectedSize:
          type: ["string", "null"]
        selectedColor:
          type: ["string", "null"]
        quantity:
          type: integer
        unitPrice:
          type: number
        totalPrice:
          type: number
        status:
          $ref: "#/components/schemas/OrderStatus"
        notes:
          type: ["string", "null"]
        confirmedAt:
          type: ["string", "null"]
        shippedAt:
          type: ["string", "null"]
        deliveredAt:
          type: ["string", "null"]
        returnedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    OrderInput:
      type: object
      required: [landingPageId, customerName, customerPhone, customerCity, deliveryZoneId, deliveryCommuneName, deliveryMethod, quantity]
      properties:
        landingPageId:
          type: integer
        customerName:
          type: string
        customerPhone:
          type: string
        customerCity:
          type: string
        customerAddress:
          type: string
        deliveryZoneId:
          type: integer
        deliveryCommuneName:
          type: string
        deliveryMethod:
          $ref: "#/components/schemas/DeliveryMethod"
        selectedSize:
          type: string
        selectedColor:
          type: string
        quantity:
          type: integer
        notes:
          type: string
    OrderUpdate:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/OrderStatus"
        notes:
          type: ["string", "null"]
        customerPhone:
          type: string
        customerCity:
          type: string
        customerAddress:
          type: ["string", "null"]
...
    OrdersPage:
      type: object
      required: [orders, total, page, limit]
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    OrdersSummary:
      type: object
      required: [total, new, pendingConfirmation, confirmed, shipped, delivered, returned, cancelled, rejected]
      properties:
        total:
          type: integer
        new:
          type: integer
        pendingConfirmation:
          type: integer
        confirmed:
          type: integer
        shipped:
          type: integer
        delivered:
          type: integer
        returned:
          type: integer
        cancelled:
          type: integer
        rejected:
          type: integer
        totalRevenue:
          type: number
        deliveryRevenue:
          type: number
        returnLoss:
          type: number
        netRevenue:
          type: number

    # ─── Customer ─────────────────────────────────────────────────────────────
...